backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
httpx>=0.27.0
google-generativeai>=0.4.0
supabase>=2.3.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
feedparser>=6.0.10
python-multipart>=0.0.9
sentence-transformers>=2.5.0
pinecone>=3.0.0
numpy>=1.24.0

